# packages/arrow/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "stately-arrow",
    "description": "Arrow backend utilities for viewing data over API",
    "version": "0.3.2"
  },
  "paths": {
    "/catalogs": {
      "get": {
        "summary": "List all catalogs in `Datafusion`",
        "description": "# Errors\n- Internal server error",
        "operationId": "list_catalogs",
        "responses": {
          "200": {
            "description": "List of registered catalogs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors": {
      "get": {
        "summary": "List all available connectors",
        "description": "# Errors\n- Internal server error",
        "operationId": "list_connectors",
        "responses": {
          "200": {
            "description": "List of available connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionMetadata"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "List databases or tables/files available in a set of connectors's underlying data stores",
        "description": "# Errors\n- Connector not found\n- Internal server error",
        "operationId": "connector_list_many",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionDetailsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "List of databases or tables/files keyed by connection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionDetailsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connector not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{connector_id}": {
      "get": {
        "summary": "List databases or tables/files available in a connector's underlying data store",
        "description": "# Errors\n- Connector not found\n- Internal server error",
        "operationId": "connector_list",
        "parameters": [
          {
            "name": "connector_id",
            "in": "path",
            "description": "Connector ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Optional search term to filter results.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of databases or tables/files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSummary"
                }
              }
            }
          },
          "404": {
            "description": "Connector not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "summary": "Execute a SQL query using URL tables",
        "description": "# Errors\n- Connector not found\n- Internal server error",
        "operationId": "execute_query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Query results as Arrow IPC stream",
            "content": {
              "application/vnd.apache.arrow.stream": {}
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Connector not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "get": {
        "summary": "List all registered connections",
        "description": "# Errors\n- Internal server error",
        "operationId": "list_registered",
        "responses": {
          "200": {
            "description": "List of registered connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionMetadata"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/register/{connector_id}": {
      "get": {
        "summary": "Register a connector. Useful when federating queries since registration is lazy",
        "description": "# Errors\n- Connector not found\n- Internal server error",
        "operationId": "register",
        "parameters": [
          {
            "name": "connector_id",
            "in": "path",
            "description": "Connector ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registered Connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionMetadata"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Connector not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Standard error shape returned by handlers",
        "required": [
          "error",
          "status"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "BackendMetadata": {
        "type": "object",
        "description": "Static metadata describing a backend implementation.\n\nBackends provide this metadata to indicate their type and capabilities.\nUse the builder methods to construct instances.",
        "required": [
          "kind",
          "capabilities"
        ],
        "properties": {
          "capabilities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "description": "Capabilities this backend supports."
          },
          "kind": {
            "$ref": "#/components/schemas/ConnectionKind",
            "description": "The kind of data source this backend connects to."
          }
        }
      },
      "Capability": {
        "type": "string",
        "description": "Capabilities a connector can expose.",
        "enum": [
          "execute_sql",
          "list"
        ]
      },
...
      "ConnectionDetailsRequest": {
        "type": "object",
        "description": "Request for fetching details from multiple connectors.",
        "required": [
          "connectors"
        ],
        "properties": {
          "connectors": {
            "type": "object",
            "description": "Map of connector IDs to their search parameters.",
            "additionalProperties": {
              "$ref": "#/components/schemas/ConnectionSearchQuery"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "fail_on_error": {
            "type": "boolean",
            "description": "If true, a failure in any connector fails the entire request.\nIf false (default), failures are skipped and successful results returned."
          }
        }
      },
      "ConnectionDetailsResponse": {
        "type": "object",
        "description": "Response containing details from multiple connectors.",
        "required": [
          "connections"
        ],
        "properties": {
          "connections": {
            "type": "object",
            "description": "Map of connector IDs to their listing results.",
            "additionalProperties": {
              "$ref": "#/components/schemas/ListSummary"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
      "ConnectionKind": {
        "oneOf": [
          {
            "type": "string",
            "description": "Object storage (S3, GCS, Azure, local filesystem).",
            "enum": [
              "object_store"
            ]
          },
          {
            "type": "string",
            "description": "Relational database (`ClickHouse`, `PostgreSQL`, etc.).",
            "enum": [
              "database"
            ]
          },
          {
            "type": "object",
            "description": "Custom connector type.",
            "required": [
              "other"
            ],
            "properties": {
              "other": {
                "type": "string",
                "description": "Custom connector type."
              }
            }
          }
        ],
        "description": "The type of data source a connector connects to."
      },
      "ConnectionMetadata": {
        "type": "object",
        "description": "Runtime metadata describing a connector instance.\n\nThis combines the connector's identity with its backend metadata,\nincluding the `DataFusion` catalog it's registered under.",
        "required": [
          "id",
          "name",
          "metadata"
        ],
        "properties": {
          "catalog": {
            "type": [
              "string",
              "null"
            ],
            "description": "The `DataFusion` catalog this connector is registered under."
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for this connector."
          },
          "metadata": {
            "$ref": "#/components/schemas/BackendMetadata",
            "description": "Backend metadata (kind and capabilities)."
          },
          "name": {
            "type": "string",
            "description": "Human-readable name."
          }
        }
      },
...
      "ConnectionSearchQuery": {
        "type": "object",
        "description": "Query parameters for searching connection contents.",
        "properties": {
          "search": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional search term to filter results."
          }
        }
      },
...
      "ListSummary": {
        "oneOf": [
          {
            "type": "object",
            "description": "List of database names (for database connectors at root level).",
            "required": [
              "summary",
              "type"
            ],
            "properties": {
              "summary": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "List of database names (for database connectors at root level)."
              },
              "type": {
                "type": "string",
                "enum": [
                  "databases"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "List of tables with metadata (for database connectors within a database).",
            "required": [
              "summary",
              "type"
            ],
            "properties": {
              "summary": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TableSummary"
                },
                "description": "List of tables with metadata (for database connectors within a database)."
              },
              "type": {
                "type": "string",
                "enum": [
                  "tables"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "List of path prefixes (for object store connectors at root level).",
            "required": [
              "summary",
              "type"
            ],
            "properties": {
              "summary": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "List of path prefixes (for object store connectors at root level)."
              },
              "type": {
                "type": "string",
                "enum": [
                  "paths"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "List of files with metadata (for object store connectors within a path).",
            "required": [
              "summary",
              "type"
            ],
            "properties": {
              "summary": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TableSummary"
                },
                "description": "List of files with metadata (for object store connectors within a path)."
              },
              "type": {
                "type": "string",
                "enum": [
                  "files"
                ]
              }
            }
          }
        ],
        "description": "Summary of items available in a connector.\n\nThe variant indicates what type of items were found based on the connector\ntype and search context."
      },
...
      "QueryRequest": {
        "type": "object",
        "description": "Request to execute a SQL query.",
        "required": [
          "sql"
        ],
        "properties": {
          "connector_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "ID of the connector to use. If not provided, the query runs against\nthe session's default catalog (if supported)."
          },
          "sql": {
            "type": "string",
            "description": "SQL query to execute. Supports URL tables like `s3://bucket/path/*.parquet`."
          }
        }
      },
...
      "TableSummary": {
        "type": "object",
        "description": "Summary information about a table or file.",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Table or file name/path."
          },
          "rows": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Number of rows (if known).",
            "minimum": 0
          },
          "size_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Size in bytes (if known).",
            "minimum": 0
          }
        }
      },
...
    },
...
  },
...
}