# packages/arrow/src/types/api.ts
import type { components } from '@/generated/types';

export type ConnectionKind = components['schemas']['ConnectionKind'];
export type ConnectionMetadata = components['schemas']['ConnectionMetadata'];
export type ListSummary = components['schemas']['ListSummary'];
export type TableSummary = components['schemas']['TableSummary'];
export type QueryRequest = components['schemas']['QueryRequest'];
export type ConnectionSearchQuery = components['schemas']['ConnectionSearchQuery'];
export type ConnectionDetailsRequest = components['schemas']['ConnectionDetailsRequest'];
export type ConnectionDetailsResponse = components['schemas']['ConnectionDetailsResponse'];